Thêm chức năng sửa, xóa
Thêm chức năng sửa, xóa

+++ 03_nguyenhuudai/03_Xay_dung_website_bang_php_thuan/quanlynhanvien/xoa.php
@@ -2,8 +2,13 @@
 if(isset($_GET['MaNV'])){
 	include 'models/connect.php';
 	$manv=$_GET['MaNV'];
-	$sql="Delete From db22_nhanvien Where MaNV=$manv";
-	$result=mysqli_query($conn,$sql);
-	header('Location:quanly.php');
+	if(!$manv){
+		header('Location:error-403.php');
+	}
+	else{
+		$sql="Delete From db22_nhanvien Where MaNV=$manv";
+		$result=mysqli_query($conn,$sql);
+		header('Location:quanly.php');
+	}
 }
 ?>
